[FIX] version_modules: function relocate, security file fix, sale description
