+++ mrp_production_custom_variants_inherited_attributes/__manifest__.py
@@ -2,7 +2,7 @@
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 {
     "name": "Production inherited attributes",
-    "version": "12.0.1.0.1",
+    "version": "12.0.1.0.2",
     "license": "AGPL-3",
     "depends": [
         "mrp_hook", "mrp_scheduled_products", "product_variant_custom",
