+++ mrp_production_custom_variants_inherited_attributes/models/__init__.py
@@ -1,5 +1,4 @@
 from . import mrp_production
-from . import product_product
 from . import product_template
 from . import mrp_bom
 from . import product_attribute
