+++ product_variant_custom/models/__init__.py
@@ -1,3 +1,4 @@
 from . import product_version
 from . import product_product
 from . import product_attribute_value
+from . import product_template
