+++ product_variant_custom/models/product_template.py
@@ -1,0 +1,15 @@
+# Copyright 2021 Mikel Arregi Etxaniz - AvanzOSC
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
+from odoo import api, fields, models
+
+
+class ProductTemplate(models.Model):
+    _inherit = "product.template"
+
+    @api.multi
+    def _get_product_attributes_dict(self):
+        if not self:
+            return []
+        self.ensure_one()
+        return self.attribute_line_ids.mapped(
+            lambda x: {'attribute_id': x.attribute_id.id})
